feat: upgrade

+++ contracts/interfaces/external/IERC20.sol
@@ -15,7 +15,7 @@ interface IERC20 {
         address recipient,
         uint256 amount
     ) external returns (bool);
-    function transferToken(address _sender, address _receiver, uint _amount) external returns (bool);
+    function transferToken(address _sender, address _receiver, uint128 _amount) external returns (bool);
     function mintToken(address account, uint256 amount) external;
     function burnToken(address account, uint256 amount) external;
     function getILPHolders() external view returns (address[] memory);
